refactor(core): move layout entrypoint
Move the layout builder into its domain folder so the core root keeps only orchestration and infrastructure entry files.
